fix: create secret with kube certs and cleanup talos secret
This PR will create a new secret called `$cluster_name-ca` that holds
certs that were generated for kubernetes. Cluster API uses these certs
to create a kubeconfig for the cluster. Additionally, in exploring this,
I discovered that we should be adding owners to our secrets so that they
get deleted automatically upon cluster deletion.

Signed-off-by: Spencer Smith <robertspencersmith@gmail.com>