feat: upgrade to CAPI v1.12.x
Upgrading the cluster-api version from v1.11.1 to v1.12.2.

Upgrading the `cluster-api` version from `v1.10.4` to `v1.11.1` with
underlying changes to Talos specific configs, same logic as before.

Inject infrastructure references for integrations tests

Due to CAPI changes, it is required to set at least one reference for
clusters and machines.
Using DevCluster (in-memory backend) and DevMachine is enough for the
integration tests.

Co-authored-by: Steffen Karlsson <steffen.karlsson@gmail.com>
Co-authored-by: Elias Werberich <elias@werberich.de>
Signed-off-by: Elias Werberich <elias@werberich.de>
Signed-off-by: Steffen Karlsson <steffen.karlsson@gmail.com>
Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>